Update to master hauls file and creates hauls file with consistent sampling area
In script 5.4 the haul file is updated with sediment and climate
variables. Also, a second file is made of hauls that is trimmed to
consistent sampling regions within each survey. This second file is to
be used with calculating annual mean clue values for gams. Script 5.5
is still being cleaned up and worked on.